Add fragment_movies_list and fragments to MainActivity
