+++ app/src/main/java/ru/flethy/androidacademyassignments/FragmentMoviesDetails.kt
@@ -1,0 +1,32 @@
+package ru.flethy.androidacademyassignments
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+
+class FragmentMoviesDetails : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+
+    override fun onCreateView(
+            inflater: LayoutInflater, container: ViewGroup?,
+            savedInstanceState: Bundle?
+    ): View? {
+        val view = inflater.inflate(R.layout.fragment_movies_details, container, false)
+
+        view?.findViewById<TextView>(R.id.back)
+                ?.apply {
+                    setOnClickListener {
+                        fragmentManager?.popBackStack()
+                    }
+                }
+
+        return view
+    }
+
+}
