--- app/src/main/java/ru/flethy/androidacademyassignments/MovieDetailsActivity.kt
@@ -1,11 +1,0 @@
-package ru.flethy.androidacademyassignments
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class MovieDetailsActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_movie_details)
-    }
-}
