+++ app/src/main/res/drawable/age_window_for_poster.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="@color/background" />
+    <corners android:radius="5dp" />
+    <size android:height="24dp" android:width="24dp" />
+</shape>
