+++ app/src/main/res/drawable/ic_target.xml
@@ -1,0 +1,22 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:aapt="http://schemas.android.com/aapt"
+    android:width="16dp"
+    android:height="16dp"
+    android:viewportWidth="16"
+    android:viewportHeight="16">
+  <path
+      android:pathData="M8.0778,0C8.2986,0 8.4823,0.1572 8.5204,0.3646L8.5277,0.4444V2.2222C8.5277,2.2278 8.5276,2.2335 8.5274,2.239C11.399,2.4547 13.6911,4.7188 13.9096,7.5556H15.5501C15.7986,7.5556 16,7.757 16,8.0055C16,8.2263 15.8408,8.41 15.631,8.4481L15.5501,8.4554H13.9087C13.6852,11.2874 11.3947,13.5462 8.5262,13.7611L8.5277,13.7778V15.5556C8.5277,15.801 8.3263,16 8.0778,16C7.8569,16 7.6732,15.8428 7.6351,15.6354L7.6279,15.5556V13.7778C7.6279,13.7721 7.628,13.7665 7.6282,13.761C4.7602,13.5456 2.4703,11.287 2.2468,8.4554H0.4499C0.2014,8.4554 0,8.254 0,8.0055C0,7.7846 0.1592,7.6009 0.369,7.5628L0.4499,7.5556H2.246C2.4644,4.7184 4.7571,2.4541 7.6294,2.239L7.6279,2.2222V0.4444C7.6279,0.199 7.8293,0 8.0778,0ZM8.0778,3.1111C10.811,3.1111 13.0268,5.2999 13.0268,8C13.0268,10.7001 10.811,12.8889 8.0778,12.8889C5.3445,12.8889 3.1287,10.7001 3.1287,8C3.1287,5.2999 5.3445,3.1111 8.0778,3.1111ZM8.0778,4.8889C6.3384,4.8889 4.9284,6.2818 4.9284,8C4.9284,9.7182 6.3384,11.1111 8.0778,11.1111C9.8171,11.1111 11.2272,9.7182 11.2272,8C11.2272,6.2818 9.8171,4.8889 8.0778,4.8889ZM8.0778,5.7778C9.3202,5.7778 10.3273,6.7727 10.3273,8C10.3273,9.2273 9.3202,10.2222 8.0778,10.2222C6.8354,10.2222 5.8282,9.2273 5.8282,8C5.8282,6.7727 6.8354,5.7778 8.0778,5.7778Z"
+      android:fillType="evenOdd">
+    <aapt:attr name="android:fillColor">
+      <gradient 
+          android:startY="5.50934"
+          android:startX="-5.25407"
+          android:endY="16.4314"
+          android:endX="5.33782"
+          android:type="linear">
+        <item android:offset="0" android:color="#FFFF4D79"/>
+        <item android:offset="1" android:color="#FFFF3365"/>
+      </gradient>
+    </aapt:attr>
+  </path>
+</vector>
