+++ app/src/main/res/values/dimens.xml
@@ -1,21 +1,35 @@
 <resources>
+    <dimen name="margin_12">12dp</dimen>
     <dimen name="margin_16">16dp</dimen>
-    <dimen name="star_margin">3dp</dimen>
+    <dimen name="star_margin">2dp</dimen>
     <dimen name="actor_margin">8dp</dimen>
     <dimen name="actor_name_margin">6dp</dimen>
     <dimen name="movie_age_margin">12dp</dimen>
-    <dimen name="back_line_margin">50dp</dimen>
+    <dimen name="movie_genre_for_poster_margin">4dp</dimen>
+    <dimen name="content_margin_top">55dp</dimen>
     <dimen name="back_icon_margin">8dp</dimen>
     <dimen name="star_line_margin">8dp</dimen>
     <dimen name="review_margin">9dp</dimen>
+    <dimen name="review_margin_for_poster">6dp</dimen>
     <dimen name="block_title_margin">24dp</dimen>
     <dimen name="actor_line_margin">6dp</dimen>
+
+    <dimen name="list_title_margin">4dp</dimen>
+    <dimen name="poster_content_margin">8dp</dimen>
+    <dimen name="star_height_for_poster">8dp</dimen>
+
+    <dimen name="title_icon_margin">8dp</dimen>
+
+    <dimen name="movies_list_title">14sp</dimen>
+    <dimen name="movie_duration">8sp</dimen>
 
     <dimen name="back">12sp</dimen>
     <dimen name="movie_name">40sp</dimen>
     <dimen name="movie_age">12sp</dimen>
     <dimen name="movie_genre">14sp</dimen>
+    <dimen name="movie_genre_for_poster">8sp</dimen>
     <dimen name="movie_review">14sp</dimen>
+    <dimen name="movie_review_for_poster">8sp</dimen>
     <dimen name="movie_storyline">14sp</dimen>
     <dimen name="block_title">14sp</dimen>
     <dimen name="actor_name">14sp</dimen>
