+++ app/src/main/res/values/strings.xml
@@ -1,27 +1,22 @@
 <resources>
     <string name="app_name">AndroidAcademyAssignments</string>
-    <string name="action_settings">Settings</string>
-    <!-- Strings used for fragments for navigation -->
-    <string name="first_fragment_label">First Fragment</string>
-    <string name="second_fragment_label">Second Fragment</string>
-    <string name="next">Next</string>
-    <string name="previous">Previous</string>
 
-    <string name="hello_first_fragment">Hello first fragment</string>
-    <string name="hello_second_fragment">Hello second fragment. Arg: %1$s</string>
+    <string name="movies_list_title">Movies list</string>
+    <string name="movie_storyline">Storyline</string>
+    <string name="movie_cast">Cast</string>
+    <string name="back">Back</string>
 
+    <!-- Avengers: End Game -->
     <string name="avengers_movie_name">Avengers: End Game</string>
     <string name="avengers_movie_age">13+</string>
+    <string name="avengers_movie_duration">137 min</string>
     <string name="avengers_movie_genre">Action, Adventure, Fantasy</string>
     <string name="avengers_movie_review">125 Reviews</string>
-    <string name="movie_storyline">Storyline</string>
     <string name="avengers_movie_storyline">After the devastating events of Avengers: Infinity War, the universe is in ruins. With the help of remaining allies, the Avengers assemble once more in order to reverse Thanos\' actions and restore balance to the universe.</string>
-    <string name="movie_cast">Cast</string>
 
     <string name="actor_1">Robert Downey Jr.</string>
     <string name="actor_2">Chris Evans</string>
     <string name="actor_3">Mark Ruffalo</string>
     <string name="actor_4">Chris Hemsworth</string>
 
-    <string name="back">Back</string>
 </resources>
