+++ app/src/main/res/values/styles.xml
@@ -1,6 +1,34 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
+    <!-- Movies list -->
+    <style name="movies_list_title">
+        <item name="fontFamily">@font/roboto_bold</item>
+        <item name="android:textSize">@dimen/movies_list_title</item>
+        <item name="android:textColor">@color/white</item>
+    </style>
+
+    <style name="movie_duration_for_poster">
+        <item name="fontFamily">@font/roboto_bold</item>
+        <item name="android:textSize">@dimen/movie_duration</item>
+        <item name="android:textColor">@color/movie_duration</item>
+        <item name="textAllCaps">true</item>
+    </style>
+
+    <style name="movie_review_for_poster">
+        <item name="fontFamily">@font/roboto_bold</item>
+        <item name="android:textSize">@dimen/movie_review_for_poster</item>
+        <item name="android:textColor">@color/movie_review</item>
+        <item name="textAllCaps">true</item>
+    </style>
+
+    <style name="movie_genre_for_poster">
+        <item name="fontFamily">@font/roboto_thin</item>
+        <item name="android:textSize">@dimen/movie_genre_for_poster</item>
+        <item name="android:textColor">@color/movie_genre</item>
+    </style>
+
+    <!-- Movies details -->
     <style name="back">
         <item name="fontFamily">@font/roboto_thin</item>
         <item name="android:textSize">@dimen/back</item>
@@ -10,19 +38,19 @@
     <style name="movie_name">
         <item name="fontFamily">@font/roboto_bold</item>
         <item name="android:textSize">@dimen/movie_name</item>
-        <item name="android:textColor">@color/white_text</item>
+        <item name="android:textColor">@color/dark_white</item>
     </style>
 
     <style name="block_title">
         <item name="fontFamily">@font/roboto_bold</item>
         <item name="android:textSize">@dimen/block_title</item>
-        <item name="android:textColor">@color/white_text</item>
+        <item name="android:textColor">@color/dark_white</item>
     </style>
 
     <style name="movie_age">
         <item name="fontFamily">@font/roboto_bold</item>
         <item name="android:textSize">@dimen/movie_age</item>
-        <item name="android:textColor">@color/white_text</item>
+        <item name="android:textColor">@color/white</item>
     </style>
 
     <style name="movie_genre">
