+++ app/src/main/res/values/themes.xml
@@ -7,7 +7,7 @@
     </style>
 
     <style name="AppTheme" parent="Theme.MaterialComponents.DayNight.NoActionBar">
-        <item name="android:statusBarColor">@color/movie_details_background</item>
+        <item name="android:statusBarColor">@color/background</item>
         <item name="windowActionBar">false</item>
         <item name="windowNoTitle">true</item>
     </style>
